Add a `Config` to thread local storage
Hang functions off of `Config` to pull `Region` or table name out of the
thread local `Config` value. These functions clone the values that
reside in thread local storage and return the clones. This is
significant in that it means for memory constrained situations these
functions should be called as infrequently as possible.