Use `KeyConstruct` to manage keys consistently
The `KeyConstruct` enables creating and granting access to KMS keys the
same across multiple constructs. Refactor the constructs to lean into
this and lift the outputs up to the `ScalableEmail` stack level rather
than defining them within each `Construct`. This allows the
`ScalableEmail` stack to be responsible for coordinating across
`Construct`s while each `Construct` focuses on the resources to do its
job.