Replace `slog` with `tracing`
The [`tracing`](https://tracing.rs) library ends up producing a
significantly smaller binary than the `slog` library. `tracing` uses
spans to create context for each event. The dependence on context was
something that `slog` eliminated by being explicit about the logger to
which a message was passed. However, `tracing` has a better integration
with `log`, which is the defacto standard for library logging in Rust,
and allows flexible usage of the emitted events and spans. Despite
current using being for structured logging it is reasonable to think the
same setup could be used for more detailed traces of data passing
through the system.